Improves feedback component and state management
Refactors feedback implementation for clarity and consistency.

Moves initial state to a dedicated file and adjusts the updateFeedback action to simplify logic.

Improves the feedback component to display relevant messages based on state.
