+++ gl-core/actions/updateFeedback.tsx
@@ -13,13 +13,7 @@ import { TFeedback } from '../types';
 export const updateFeedback =
   (feedback: TFeedback | null) => async (dispatch: TUbereduxDispatch) => {
     try {
-      console.log('updateFeedback', feedback);
-
-      const updated = {
-        ...(feedback || {}),
-        hidden: false,
-      };
-
+      const updated = {...(feedback || {})};
       dispatch(
         setUbereduxKey({
           key: 'feedback',
